Renamed android bundle and applied AS lint fixes.
