+++ android/build.gradle
@@ -88,7 +88,7 @@ def configureReactNativePom(def pom) {
         name packageJson.title
         artifactId packageJson.name
         version = packageJson.version
-        group = "com.reactlibrary"
+        group = "com.sudoplz.rninappupdates"
         description packageJson.description
         url packageJson.repository.baseUrl
